requirements.txt
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.5
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.15.0
numpy==2.4.1
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pillow==12.1.0
protobuf==6.33.4
pyarrow==23.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
smmap==5.0.2
speedtest-cli==2.1.3
streamlit==1.53.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
